+++ shoop/admin/modules/categories/__init__.py
@@ -10,6 +10,7 @@ from django.db.models import Q
 from django.utils.translation import ugettext_lazy as _
 
 from shoop.admin.base import AdminModule, MenuEntry, SearchResult
+from shoop.admin.utils.permissions import get_default_model_permissions
 from shoop.admin.utils.urls import (
     derive_model_url, get_edit_and_list_urls, get_model_url
 )
@@ -25,7 +26,8 @@ class CategoryModule(AdminModule):
         return get_edit_and_list_urls(
             url_prefix="^categories",
             view_template="shoop.admin.modules.categories.views.Category%sView",
-            name_template="category.%s"
+            name_template="category.%s",
+            permissions=get_default_model_permissions(Category),
         )
 
     def get_menu_entries(self, request):
@@ -51,6 +53,9 @@ class CategoryModule(AdminModule):
                     category=self.category,
                     relevance=relevance
                 )
+
+    def get_required_permissions(self):
+        return get_default_model_permissions(Category)
 
     def get_model_url(self, object, kind):
         return derive_model_url(Category, "shoop_admin:category", object, kind)