+++ shoop/admin/modules/services/__init__.py
@@ -11,6 +11,7 @@ from __future__ import unicode_literals
 from django.utils.translation import ugettext_lazy as _
 
 from shoop.admin.base import AdminModule, MenuEntry
+from shoop.admin.utils.permissions import get_default_model_permissions
 from shoop.admin.utils.urls import (
     admin_url, derive_model_url, get_edit_and_list_urls
 )
@@ -28,16 +29,19 @@ class ServiceModule(AdminModule):
     url_name_prefix = None
 
     def get_urls(self):
+        permissions = self.get_required_permissions()
         return [
             admin_url(
                 "%s/(?P<pk>\d+)/delete/$" % self.url_prefix,
                 self.view_template % "Delete",
-                name=self.name_template % "delete"
+                name=self.name_template % "delete",
+                permissions=permissions
             )
         ] + get_edit_and_list_urls(
             url_prefix=self.url_prefix,
             view_template=self.view_template,
-            name_template=self.name_template
+            name_template=self.name_template,
+            permissions=permissions
         )
 
     def get_menu_entries(self, request):
@@ -48,6 +52,9 @@ class ServiceModule(AdminModule):
                 category=self.category
             )
         ]
+
+    def get_required_permissions(self):
+        return get_default_model_permissions(PaymentMethod) | get_default_model_permissions(ShippingMethod)
 
     def get_model_url(self, object, kind):
         return derive_model_url(self.model, self.url_name_prefix, object, kind)